Fix ln/log evaluation, savings term math, and UI efficiency
- Map keypad ln/log to exp4j log/log10; fix e) corruption in ln(e)
- Substitute standalone Euler e; align savings months with loan (fractional years)
- Currency screen uses warm cache (no loading flash on revisit)
- Main calculator: single formatExpression pass, safer history restore
- Guard sales tax against negative price or rate

+++ app/src/main/java/com/techtree/clevcaleb/logic/Calculations.kt
@@ -25,8 +25,8 @@ object FinanceCalculations {
         years: Double,
         taxOnInterestPercent: Double,
     ): Triple<Double, Double, Double>? {
-        val months = (years * 12).toInt()
-        if (months <= 0) return null
+        val months = years * 12
+        if (months <= 0.0) return null
         val r = annualRatePercent / 100 / 12 * (1 - taxOnInterestPercent / 100)
         val balance = if (r == 0.0) {
             initial + monthly * months
